Change Overlay to be a hidden component (#693)
