--- core/components/atoms/overlay/overlay.md
@@ -1,18 +1,0 @@
-```meta
-  category: Layout
-  description: Low-level component that can be used to render any content on top of the regular viewport
-```
-
-`import { Overlay } from '@auth0/cosmos'`
-
----
-
-You probably won't need to use it directly; In most cases a <code>Dialog</code> might be more appropriate.
-
-```jsx
-<Overlay onClose={() => {}}>Hello from Cosmos!</Overlay>
-```
-
-Here's an example <code>Overlay</code>:
-
-![overlay example](screenshots/overlay.png)
