--- core/components/atoms/overlay/overlay.sketch.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { storiesOf } from '@storybook/react'
-
-import { Overlay } from '@auth0/cosmos'
-
-storiesOf('Overlay', module).add('open', () => (
-  <Overlay open onClose={() => {}}>
-    Hello from Cosmos!
-  </Overlay>
-))
