+++ core/components/index.js
@@ -20,7 +20,6 @@ import {
   Link,
   List,
   Logo,
-  Overlay,
   PageHeader,
   Pagination,
   PaginationToolbar,
@@ -64,7 +63,6 @@ export {
   Link,
   List,
   Logo,
-  Overlay,
   PageHeader,
   Pagination,
   PaginationToolbar,
