+++ core/components/molecules/dialog/dialog.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import Button from '../../atoms/button'
 import ButtonGroup from '../../molecules/button-group'
-import Overlay from '../../atoms/overlay'
+import Overlay from '../../atoms/_overlay'
 import Icon from '../../atoms/icon'
 import Link from '../../atoms/link'
 import DialogAction from './dialog-action'
